@@ -1,11 +1,11 @@
 name: flutter_dropdown_alert
 description: A dropdown alert package for flutter, help user notify the alert like push notification when call api success, fail or waring.
-version: 0.0.9
+version: 1.0.0
 email: vantuan88291@gmail.com
 homepage: https://github.com/vantuan88291/flutter_dropdown_alert
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
