@@ -2,7 +2,7 @@ class DataAlert {
   String message;
   String title;
   TypeAlert type;
-  Map<String, dynamic> payload;
+  Map<String, dynamic>? payload;
 
   DataAlert(this.message, this.title, this.type, {this.payload});
 }
